Update to protocol 12.61 (#2302)
- [X] Fixed Appveyor warnings (Thanks for @marcosvf132 for help)
- [X] Added new items (added in last 12.60 version)

+++ src/player.h
@@ -414,7 +414,7 @@ class Player final : public Creature, public Cylinder
 			operatingSystem = clientos;
 		}
 
-		uint16_t getProtocolVersion() const {
+		uint32_t getProtocolVersion() const {
 			if (!client) {
 				return 0;
 			}
